discorddb/pyproject.toml
[tool.poetry]
name = "discorddb"
version = "0.1.0"
description = ""
authors = ["nx1p <email@marci.tech>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
"discord.py" = "^2.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
